commit -> WeatherApi Call and setting data done.

+++ app/src/main/java/com/tufelmalik/dailykill/data/model/weather/Wind.kt
@@ -2,6 +2,6 @@ package com.tufelmalik.dailykill.data.model.weather
 
 data class Wind(
     val deg: Int,
-    val gust: Int,
+    val gust: Double,
     val speed: Double
 )
